+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # vdlx-datagrid changelog
 ## v4.8.1
 
+- Columns automatically switch to read-only when the bound scenario is reserved for execution. 
+
 ## v4.8.0
+
 - Updated to VDL 4.8 compatibility.
 
 ## v4.7.2